Copied and altered
I copied Pahimar's LetsModRebbot git, edited the files to match my build
structure, and created the portal block, and the icon for creative tab.

Portal block acts like cobweb when you enter, and I plan on generating
pockets of it in the nether to transport players to Second Circle.

The portal block is unbreakable and uncraftable. It does not generate
with any player built structure. Found in environment only.