Fixed portal
I fixed the portal from the nether to Second Circle dim. Igniter
craftable with nether star. Igniter can only be used in the Nether or
SCDim. Working on actual portal frame structure still, but working with
a 5x5x5 frame idea. currently the portal frame is set to just the base.
ANNNA
NNNNN
NNNNN
NNNNN
ANNNA
A = Anything
N = Nether Brick
Use igniter in the center block.

I plan on adding pillars to the corners, and a roof.